created a locktype file
