+++ src/data/LockConstants.ts
@@ -1,0 +1,13 @@
+export const YEARS_LOCK = 4
+export const YIELD_GAS_LIMIT = 185000
+export const CHECKPOINT_GAS_LIMIT = 220000
+export const LOCK_UPDATE_GAS_LIMIT = 440000
+export const LOCK_AND_WITHDRAWAL_GAS_LIMIT = 490000
+export const DEFAULT_GAS_LIMIT = 170000
+export const APPROVE = 50000
+export const EP_COINGECKO_URL =
+  'https://api.coingecko.com/api/v3/simple/price?ids=everipedia&vs_currencies=usd'
+export const IQ_TOKEN_HOLDER =
+  'https://ethplorer.io/service/service.php?data=0x1bf5457ecaa14ff63cc89efd560e251e814e16ba&page=pageSize%3D600%26pageTab%3Dholders%26tab%3Dtab-holders&showTx=all'
+export const NORMALIZE_VALUE = 10e17
+export const REWARDS_FOR_THE_FIRST_YEAR = 1095000000 // 3M per day for 365 days
